split up coconut module a bit
